Tweak generation of ToolsUI JNLP files so they work on our web server
ui.jar's 'Codebase' manifest attribute and the /jnlp/@codebase attributes of ToolsUI JNLP files are now kept in sync. They are controlled by a single variable in ':ui': webstartCodebase. It turns out that different values for all of these codebases was the principal cause of the "Application Blocked by Java Security" failures we were seeing in javaws.
* The ToolsUiJnlp tasks now have a codebase property. Updated unit tests to exercise the new property.
* The aforementioned property may be 'null', which suppresses output of the /jnlp/@codebase attribute. This is useful for local Web Start testing.
* Restored my (now simpler) notes about how to test Web Start locally.

Other changes:
* Don't include /jnlp/@href attribute in ToolsUiJnlp*Task output files.
* ToolsUiJnlpExtensionTask outputs must include '<security><all-permissions/></security>' element.
* Refactored ToolsUiJnlp*Task tests to be simpler and more consistent.
* For all URLs in our JAR manifests and JNLP files that point to our websever, use HTTPS instead of HTTP.